pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-srt-translator"
version = "2.1.4"
description = "A tool to translate subtitles using Google Generative AI"
readme = "README.md"
license = "MIT"
authors = [{ name = "Matheus Castro" }]
dependencies = [
  "google-genai==1.23.0",
  "srt==3.5.3",
  "json-repair==0.47.6",
  "packaging==25.0",
  "pysubs2==1.6.1",
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.10"

[project.scripts]
gemini-srt-translator = "gemini_srt_translator.cli:main"
gst = "gemini_srt_translator.cli:main"
